Added social icons (but with no links), and starting to learn about forms.
Forms are well-suited for users browsing entrepreneurial/business websites, but I want to see if they work for smaller websites as well. Since this website is hosted on Github, I have no way of knowing what will happen if I submit the content. It probably doesn't work.

I took a long break, and will probably continue to take it easy for a while. I might do some cleanup to make the HTML readable, with proper indents and all on all of the pages.